requirements.txt
django==4.0.5
pillow==9.1.1
django-rosetta==0.9.8
django-crispy-forms==1.14.0
django-filter==22.1